requirements.txt
python-dotenv
google-genai
python-dotenv